refactor: Removed ICallableGenerator and it's usages to use the same method for all different types of Federations on ServiceMethodHelper.cs
